refactor(rag_core): extract ingestion + index I/O to rag_indexing (closes #858)
ADR 0045 G3 — second large extraction in the rag_core slim-down series after rag_embedding (G2 / PR #847). Pure structural relocation.

New leaf module rag_indexing.py owns the raw-document → chunk → index-payload pipeline plus the index I/O layer:

  - load_raw_documents, normalize_json_document, normalize_text_document
  - validate_chunking_options, resolve_chunking_strategy
  - build_chunk_records, build_chunks, make_chunk
  - build_index_payload, build_index_payload_from_documents
  - load_index, write_index
  - known_entities, metadata_targets
  - constants: DEFAULT_CHUNK_MAX_CHARS, DEFAULT_CHUNK_OVERLAP_SENTENCES, VALID_CHUNKING_STRATEGIES, INDEX_FILENAME, INDEX_SCHEMA_VERSION

Dependencies: stdlib (json), numpy, rag_text_processing (tokenize, normalize_entity), rag_metadata_processing (normalize_regions, normalize_page_span, normalize_document_sections, fixed_parent_section, split_section_text, make_metadata_target), rag_embedding (DEFAULT_EMBEDDING_MODEL, embed_texts), rag_vector_store (vector_store_from_matrix, load_vector_store, EMBEDDINGS_FILENAME). Zero back-edges to rag_core / rag_retrieval / rag_query / rag_verifier / rag_answer — a true leaf.

rag_core re-exports every public symbol so existing call sites (from rag_core import build_index_payload, load_index, DEFAULT_CHUNK_MAX_CHARS, etc.) keep working unchanged.

Verification (ADR 0045):
  - 233/234 pytest cases pass under `tests/ -k "naive_baseline or retrieval or indexing or ingestion or vector_store or chunk"`. The single failure (test_mixed_format_ingestion_regression::test_text_source_counts_aggregated_by_format) is pre-existing on origin/main HEAD — verified by replacing ingestion.py + the test with their main versions and reproducing the same failure. Unrelated to this PR; tracked separately.
  - ADR 0001 naive_baseline preset golden unchanged
  - rag_core.py shrunk 1628 → 1263 LOC (-365)
  - rag_indexing.py new at 467 LOC

Out of scope (follow-ups):
  - G4: dependency-graph final verification + re-export shim removal after consumer audit
  - Test fix for the pre-existing fallback_reasons assertion drift (separate concern)
